move edit asset files/folders
