+++ src/pages/CreateOrEditAssetPage/EditAssetForm/AssetSummary.vue
