+++ src/pages/CreateOrEditAssetPage/EditWidget/EditDateWidget.vue
@@ -27,12 +27,12 @@
     "
     @update:widgetContents="
       (widgetContents) => {
-        $emit('update:widgetContents', widgetContents as Type.WithId<Type.LocationWidgetContent>[]);
+        $emit('update:widgetContents', widgetContents as Type.WithId<Type.DateWidgetContent>[]);
       }
     ">
     <template #fieldContents="{ item }">
-      <EditLocationWidgetContentItem
-        :modelValue="(item as Type.WithId<Type.LocationWidgetContent>)"
+      <EditDateWidgetContentItem
+        :modelValue="(item as Type.WithId<Type.DateWidgetContent>)"
         @update:modelValue="handleItemUpdate" />
     </template>
   </EditWidgetLayout>
@@ -40,40 +40,37 @@
 <script setup lang="ts">
 import * as Type from "@/types";
 import EditWidgetLayout from "./EditWidgetLayout.vue";
-import * as ops from "../editWidgetOps";
-import EditLocationWidgetContentItem from "./EditLocationWidgetContentItem.vue";
+import * as ops from "./helpers/editWidgetOps";
+import EditDateWidgetContentItem from "./EditDateWidgetContentItem.vue";
 
 const props = defineProps<{
-  widgetDef: Type.LocationWidgetDef;
-  widgetContents: Type.WithId<Type.LocationWidgetContent>[];
+  widgetDef: Type.DateWidgetDef;
+  widgetContents: Type.WithId<Type.DateWidgetContent>[];
   isOpen: boolean;
 }>();
 
 const emit = defineEmits<{
   (
     e: "update:widgetContents",
-    widgetContents: Type.WithId<Type.LocationWidgetContent>[]
+    widgetContents: Type.WithId<Type.DateWidgetContent>[]
   ): void;
   (e: "update:isOpen", isOpen: boolean): void;
 }>();
 
-function handleItemUpdate(
-  updatedItem: Type.WithId<Type.LocationWidgetContent>
-) {
+function handleItemUpdate(updatedItem: Type.WithId<Type.DateWidgetContent>) {
   const index = props.widgetContents.findIndex((i) => i.id === updatedItem.id);
   if (index === -1) {
     throw Error(
-      `Cannot update location widget: item with id "${updatedItem.id}" not found`
+      `Cannot update date widget: item with id "${updatedItem.id}" not found`
     );
   }
 
-  const updatedContents = [
+  const updatedWidgetContents = [
     ...props.widgetContents.slice(0, index),
     updatedItem,
     ...props.widgetContents.slice(index + 1),
   ];
-
-  emit("update:widgetContents", updatedContents);
+  emit("update:widgetContents", updatedWidgetContents);
 }
 </script>
 <style scoped></style>
