+++ src/pages/CreateOrEditAssetPage/EditWidget/EditLocationWidgetContentItem.vue
