+++ src/pages/CreateOrEditAssetPage/EditWidget/EditTagWidget.vue
@@ -38,7 +38,7 @@
 <script setup lang="ts">
 import * as Type from "@/types";
 import EditWidgetLayout from "./EditWidgetLayout.vue";
-import * as ops from "../editWidgetOps";
+import * as ops from "./helpers/editWidgetOps";
 import {
   TagsInput,
   TagsInputItem,
