+++ src/pages/CreateOrEditAssetPage/EditWidget/EditWidget.vue
