+++ src/pages/CreateOrEditAssetPage/EditWidget/FileUploader.vue
