+++ src/pages/CreateOrEditAssetPage/EditWidget/helpers/findDeepestPath.test.ts
