+++ src/pages/CreateOrEditAssetPage/EditWidget/helpers/findDeepestPath.ts
