+++ src/pages/CreateOrEditAssetPage/TableOfContents/TableOfContents.vue
